// CallCenterAngular/src/app/models/st-basic-vm.model.ts
export interface StBasicVm {
  sIdno: string;
  brokerNo4: string;
  sAcno: string;
  sName: string;
  nbankName: string;
  nbankAcctNo: string;
  obankName: string;
  obankAcctNo: string;
  investorSex: string;
  birthday: string;
  zipBirth: string;
  zipCont: string;
  addrBirth: string;
  addrCont: string;
  investorTel1: string;
  investorTel2: string;
  investorMobile: string;
  saleCode: string;
  freezeReason: string;
  tradeStatus: string;
  legalAgent1Name: string;
  agentName: string;
  agentId: string;
  agentBirthday: string;
  agentCloseDate: string;
  bondAcctNo: string;
  acctCloseDate: string;
  tradeLimit: string;
  mmaAcctNo: string;
  personalData: string;
  referencesName: string;
  referencesTel: string;
  promoBank: string;
  promoFutures: string;
  financialAdvisor: string;
  legalAgent1Id: string;
  legalAgent1Birthday: string;
  tel21: string;
  earmarkEndDate: string;
  companyName: string;
  emergingRisk: string;
  emergingRiskNew: string;
  firstEmerging: string;
  warrantRisk: string;
  otcRisk: string;
  warrantRiskNew: string;
  cbbcRisk: string;
  foreignSecRisk: string;
  riskDisclosureStatement: string;
  firstSecRisk: string;
  japanNote: string;
  futuresEtfRisk: string;
  ultrashortEtf: string;
  ultrashortEtfTrade: string;
  spotGold: string;
  mediaNetReceipts: string;
  legalAgent2Name: string;
  legalAgent2Id: string;
  legalAgent2Birthday: string;
  loanContract: string;
  loanRates: string;
  daytradeRiskAgree: string;
  dayTradeDate: string;
  dayTradeLimit: string;
  dayTradeStatus: string;
  acctCreateDate: string;
  promoInvest: string;
  earmarkStartDate: string;
  investorEmail: string;
  acctSource: string;
  career: string;
  launderRisk: string;
  amlRisk: string;
  officeTel1: string;
  officeTel2: string;
  currencyEtfRisk: string;
  pBankBookStartDate: string;
  pBankBookEndDate: string;
  phoneLastDate: string;
  onlineTradingContract: string;
  earMark: string;
  arMarkAmt: string;
  etf01: string;
  postWay: string;
  wmDate: string;
  riskDisclosureWithFu: string;
  etnStatement: string;
  etnTrade: string;
  transStatement: string;
  etnNotToSignDate: string;
  promoRest: string;
  promoInvestments: string;
  sameIpStatement: string;
  sameIpDate: string;
  cuaey: string;
  cuaez: string;
  cuae0: string;
  cuaf0Date: string;
  etnRiskStrategyStatement: string;
  fuEtfTest: string;
  investorFax: string;
  fafApplyDate: string;
  fafCancelDate: string;
  spOrg: string;
}